requirements.txt
python-dotenv
discord
pypresence
flask
aiohttp
requests